Fix calibration test on mac
