+++ testing/calibration/contract/fil-gas-calibration-actor/Cargo.toml
@@ -4,10 +4,12 @@ version = "0.1.0"
 edition = "2021"
 publish = false
 
-[dependencies]
+[target.'cfg(target_arch = "wasm32")'.dependencies]
 fvm_sdk = { version = "3.0.0", path = "../../../../sdk" }
 fvm_shared = { version = "3.1.0", path = "../../../../shared" }
 fvm_ipld_encoding = { version = "0.3.3", path = "../../../../ipld/encoding" }
+fvm_gas_calibration_shared = { path = "../../shared" }
+
 cid = { version = "0.8.2", default-features = false }
 num-derive = "0.3"
 num-traits = "0.2"
@@ -15,4 +17,4 @@ serde = { version = "1.0", features = ["derive"] }
 anyhow = "1.0.51"
 
 [lib]
-crate-type = ["cdylib", "lib"]
+crate-type = ["cdylib"]
