+++ testing/calibration/src/bin/on_hashing.rs
@@ -2,8 +2,8 @@
 // SPDX-License-Identifier: Apache-2.0, MIT
 #![feature(slice_group_by)]
 
-use fil_gas_calibration_actor::{Method, OnHashingParams};
 use fvm_gas_calibration::*;
+use fvm_gas_calibration_shared::{Method, OnHashingParams};
 use fvm_shared::crypto::hash::SupportedHashes;
 use rand::{thread_rng, Rng};
