+++ testing/integration/tests/fil-malformed-syscall-actor/Cargo.toml
@@ -4,7 +4,7 @@ version = "0.1.0"
 edition = "2021"
 publish = false
 
-[dependencies]
+[target.'cfg(target_arch = "wasm32")'.dependencies]
 fvm_shared = { version = "3.1.0", path = "../../../../shared" }
 fvm_sdk = { version = "3.0.0", path = "../../../../sdk" }
